finished the filterVCmod
