+++ UberEATS/Controllers/Home/Filter/FilterViewMenu.swift
@@ -10,7 +10,6 @@ import UIKit
 
 class FilterViewMenuCell: UICollectionViewCell {
     
-    
     lazy var menuLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -32,7 +31,6 @@ class FilterViewMenuCell: UICollectionViewCell {
         }
     }
     
-    
     override init(frame: CGRect) {
         super.init(frame: frame)
         setupViews()
@@ -51,7 +49,6 @@ class FilterViewMenuCell: UICollectionViewCell {
             menuLabel.bottomAnchor.constraint(equalTo: bottomAnchor, constant: 0)
             ])
     }
-    
 }
 
 class FilterViewMenu: UIView, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout, UICollectionViewDataSource {
@@ -111,6 +108,7 @@ class FilterViewMenu: UIView, UICollectionViewDelegate, UICollectionViewDelegate
             collectionView.bottomAnchor.constraint(equalTo: bottomAnchor, constant: 0)
             ])
     }
+    
     override init(frame: CGRect) {
         super.init(frame: frame)
         collectionView.selectItem(at: IndexPath.init(item: 0, section: 0), animated: true, scrollPosition: .centeredHorizontally)
@@ -120,7 +118,6 @@ class FilterViewMenu: UIView, UICollectionViewDelegate, UICollectionViewDelegate
     required init(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
 }
