+++ UberEATS/Utils/UIImageView+load.swift
@@ -32,3 +32,14 @@ extension UIImage {
     }
 }
 
+extension UIImage {
+    func alpha(_ value:CGFloat)->UIImage?
+    {
+        UIGraphicsBeginImageContextWithOptions(size, false, scale)
+        draw(at: CGPoint.zero, blendMode: .normal, alpha: value)
+        let newImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        return newImage
+    }
+}
+
